passphrase part 2 day4
build a char frequency dictionary for each word, compare it with other words of the same length.  If you find a character that's not in another word or it doesn't show up with the same freq, then it can't be an anagam.  If it is the same, the break since it's invalid.